+++ processor/transformprocessor/internal/common/expression.go
@@ -58,6 +58,10 @@ func (g exprGetter) Get(ctx TransformContext) interface{} {
 }
 
 func NewGetter(val Value, functions map[string]interface{}, pathParser PathExpressionParser) (Getter, error) {
+	if val.IsNil != nil && *val.IsNil {
+		return &literal{value: nil}, nil
+	}
+
 	if s := val.String; s != nil {
 		return &literal{value: *s}, nil
 	}